feat: 컴포넌트 분리 및 로딩 및 에러 핸들링
