+++ frontend/src/pages/Admin/EnterPhoneNumber/components/GuideAlert/style.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const BackDrop = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  z-index: 10;
+  background-color: rgba(0, 0, 0, 0.3);
+`;
