+++ frontend/src/pages/Admin/EnterPhoneNumber/index.tsx
@@ -1,21 +1,15 @@
 import Dialpad from './components/Dialpad';
 import { IoIosArrowBack } from '@react-icons/all-files/io/IoIosArrowBack';
-import {
-  Container,
-  IconWrapper,
-  PrivacyBox,
-  RowContainer,
-  TableItem,
-  TableTitleItem,
-  Title,
-} from './style';
-import { useNavigate } from 'react-router-dom';
+import { Container, IconWrapper, Title } from './style';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { useRedirectRegisterPage } from '../../../hooks/useRedirectRegisterPage';
 import PrivateProvider from '../../../provider/PrivateProvider';
 import ROUTER_PATH from '../../../constants/routerPath';
+import PrivacyBox from './components/PrivacyBox';
 
 const EnterPhoneNumber = () => {
   useRedirectRegisterPage();
+  const location = useLocation();
   const navigate = useNavigate();
 
   const navigateBack = () => {
@@ -28,31 +22,10 @@ const EnterPhoneNumber = () => {
         <IconWrapper onClick={navigateBack}>
           <IoIosArrowBack size="40" />
         </IconWrapper>
-        전화번호 입력
+        {location.pathname === ROUTER_PATH.enterStamp ? '스탬프 적립' : '리워드 사용'}
       </Title>
       <Container>
-        <PrivacyBox>
-          <h1>개인정보 수집 및 이용동의</h1>
-          <p>
-            스탬프크러쉬 서비스 회원가입, 고지사항 전달 등을 위해 아래와 같이 개인정보를 수집 및
-            이용합니다. <br /> 전화번호를 입력하시면 개인정보 수집 및 이용에 동의하시는 것으로
-            간주됩니다.
-          </p>
-          <RowContainer>
-            <TableTitleItem>수집 목적</TableTitleItem>
-            <TableTitleItem>수집 항목</TableTitleItem>
-            <TableTitleItem>수집 근거</TableTitleItem>
-            <TableTitleItem>수집 기간</TableTitleItem>
-          </RowContainer>
-          <RowContainer>
-            <TableItem>
-              회원 식별 및 회원제 서비스 제공과 <br /> 서비스 변경사항 및 고지사항 전달
-            </TableItem>
-            <TableItem>전화번호</TableItem>
-            <TableItem>개인정보 보호법 제 15조 1항</TableItem>
-            <TableItem>회원탈퇴 시까지</TableItem>
-          </RowContainer>
-        </PrivacyBox>
+        <PrivacyBox />
         <Dialpad />
       </Container>
     </PrivateProvider>
